# Retrieve productVersion using informational version attribute in AmsiUtils.Init()