--- a/src/System.Management.Automation/security/SecuritySupport.cs
+++ b/src/System.Management.Automation/security/SecuritySupport.cs
@@ -856,6 +856,7 @@ namespace Microsoft.PowerShell.Commands
 
 namespace System.Management.Automation
 {
+    using System.Reflection;
     using System.Security.Cryptography.Pkcs;
 
     /// <summary>
@@ -1336,35 +1337,15 @@ namespace System.Management.Automation
 
     internal static class AmsiUtils
     {
-        private static string GetProcessHostName(string processName)
-        {
-            return string.Concat("PowerShell_", processName, ".exe_0.0.0.0");
-        }
-
         internal static int Init()
         {
             Diagnostics.Assert(s_amsiContext == IntPtr.Zero, "Init should be called just once");
 
             lock (s_amsiLockObject)
             {
-                Process currentProcess = Process.GetCurrentProcess();
-                string hostname;
-                try
-                {
-                    hostname = string.Concat("PowerShell_", Environment.ProcessPath, "_",
-                        currentProcess.MainModule.FileVersionInfo.ProductVersion);
-                }
-                catch (ComponentModel.Win32Exception)
-                {
-                    // This exception can be thrown during thread impersonation (Access Denied for process module access).
-                    hostname = GetProcessHostName(currentProcess.ProcessName);
-                }
-                catch (FileNotFoundException)
-                {
-                    // This exception can occur if the file is renamed or moved to some other folder
-                    // (This has occurred during Exchange set up).
-                    hostname = GetProcessHostName(currentProcess.ProcessName);
-                }
+                Assembly currentAssembly = typeof(AmsiUtils).Assembly;
+                string productVersion = currentAssembly.GetCustomAttribute<AssemblyInformationalVersionAttribute>().InformationalVersion;
+                string hostname = string.Concat("PowerShell_", Environment.ProcessPath, "_", productVersion);
 
                 AppDomain.CurrentDomain.ProcessExit += CurrentDomain_ProcessExit;