@iSazonov 2021-06-06
Since it is about startup performance please share perf results for just startup. Either `Measure-Command { .\pwsh -noprofile -c exit }` or PerfView.