@Fs00 2021-06-06
Here are the results in milliseconds of running `Measure-Command { .\pwsh -noprofile -c exit }` on a debug build 15 times:

**master branch:**
854,19, 807,08, 817,18, 804,88, 820,26, 810,25, 817,35, 829,89, 804,71, 804,96, 809,28, 815,76, 811,47, 810,23, 800,47
Best: 800,47 ms
Worst: 854,19 ms
Average: **814,53 ms**

**This PR:**
785,93, 788,26, 795,68, 786,38, 789,25, 792,89, 791,65, 806,51, 781,68, 791,21, 796,45, 796,87, 791,29, 785,23, 787,25
Best: 781,68 ms
Worst: 806,51 ms
Average: **791,10 ms**