@iSazonov 2021-06-06
> on a debug build 15 times:

Please do all perf test on Release ( Start-PSBuild -Configuration Release -CrossGen)