@iSazonov 2021-06-14
Not for the PR but while we are here - what do you think about using new c#  feature - source generator - for getting ProductVersion?